# Proposal: Change `Query` model