--- a/core/shared/src/main/scala/org/http4s/Query.scala
+++ b/core/shared/src/main/scala/org/http4s/Query.scala
@@ -34,83 +34,257 @@ import java.nio.charset.StandardCharsets
 import scala.collection.mutable
 import scala.collection.mutable.ListBuffer
 
-/** Collection representation of a query string
+/** Representation of a query string.
   *
-  * It is a indexed sequence of key and maybe a value pairs which maps
-  * precisely to a query string, modulo
+  * When a query is none, it is represented by the [[Query.Empty]].
+  *
+  * When a query is parsed – it is represented by the [[Query.Parsed]],
+  * an indexed sequence of key and maybe value pairs
+  * which maps precisely to a query string, modulo
   * [[https://datatracker.ietf.org/doc/html/rfc3986#section-2.1 percent-encoding]].
+  * The resulting `String` will have the pairs separated
+  * by '&' while the key is separated from the value with '='.
   *
-  * When rendered, the resulting `String` will have the pairs separated
-  * by '&' while the key is separated from the value with '='
+  * Otherwise, a query is represented by the [[Query.Raw]] containing unparsed string.
   */
-final class Query private (value: Either[Vector[KeyValue], String])
-    extends QueryOps
-    with Renderable {
-  private[this] var _pairs: Vector[KeyValue] = null
 
-  def pairs: Vector[KeyValue] = {
-    if (_pairs == null) {
-      _pairs = value.fold(identity, Query.parse)
-    }
-    _pairs
+sealed trait Query extends QueryOps with Renderable {
+  def pairs: Vector[KeyValue]
+
+  def get(idx: Int): Option[KeyValue] = this match {
+    case Query.Empty => None
+    case parsed: Query.Parsed => parsed.pairs.get(idx.toLong)
+    case raw: Query.Raw => raw.pairs.get(idx.toLong)
   }
 
-  private def this(vec: Vector[KeyValue]) = this(Left(vec))
+  def length: Int = this match {
+    case Query.Empty => 0
+    case parsed: Query.Parsed => parsed.pairs.length
+    case raw: Query.Raw => raw.pairs.length
+  }
 
-  @deprecated("Unsafe method. Use get(idx) instead", "0.23.17")
-  def apply(idx: Int): KeyValue = pairs(idx)
+  def slice(from: Int, until: Int): Query = this match {
+    case Query.Empty => this
 
-  def get(idx: Int): Option[KeyValue] = pairs.get(idx.toLong)
+    case parsed: Query.Parsed =>
+      val sliced = parsed.pairs.slice(from, until)
+      if (sliced.lengthIs == 0) Query.Empty
+      else new Query.Parsed(sliced)
 
-  def length: Int = pairs.length
+    case raw: Query.Raw =>
+      val sliced = raw.pairs.slice(from, until)
+      if (sliced.lengthIs == 0) Query.Empty
+      else new Query.Parsed(sliced)
+  }
 
-  def slice(from: Int, until: Int): Query = new Query(Left(pairs.slice(from, until)))
+  def isEmpty: Boolean = this match {
+    case Query.Empty => true
+    case parsed: Query.Parsed => parsed.pairs.isEmpty
+    case raw: Query.Raw => raw.pairs.isEmpty
+  }
 
-  def isEmpty: Boolean = pairs.isEmpty
+  def nonEmpty: Boolean = !isEmpty
 
-  def nonEmpty: Boolean = pairs.nonEmpty
+  def drop(n: Int): Query = this match {
+    case Query.Empty => this
 
-  def drop(n: Int): Query = new Query(Left(pairs.drop(n)))
+    case parsed: Query.Parsed =>
+      val prepared = parsed.pairs.drop(n)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
 
-  def dropRight(n: Int): Query = new Query(Left(pairs.dropRight(n)))
+    case raw: Query.Raw =>
+      val prepared = raw.pairs.drop(n)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
+  }
 
-  def exists(f: KeyValue => Boolean): Boolean =
-    pairs.exists(f)
+  def dropRight(n: Int): Query = this match {
+    case Query.Empty => this
 
-  def filterNot(f: KeyValue => Boolean): Query =
-    new Query(Left(pairs.filterNot(f)))
+    case parsed: Query.Parsed =>
+      val prepared = parsed.pairs.dropRight(n)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
 
-  def filter(f: KeyValue => Boolean): Query =
-    new Query(Left(pairs.filter(f)))
+    case raw: Query.Raw =>
+      val prepared = raw.pairs.dropRight(n)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
+  }
 
-  def foreach(f: KeyValue => Unit): Unit =
-    pairs.foreach(f)
+  def exists(f: KeyValue => Boolean): Boolean = this match {
+    case Query.Empty => false
+    case parsed: Query.Parsed => parsed.pairs.exists(f)
+    case raw: Query.Raw => raw.pairs.exists(f)
+  }
 
-  def foldLeft[Z](z: Z)(f: (Z, KeyValue) => Z): Z =
-    pairs.foldLeft(z)(f)
+  def filter(f: KeyValue => Boolean): Query = this match {
+    case Query.Empty => this
+
+    case parsed: Query.Parsed =>
+      val prepared = parsed.pairs.filter(f)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
+
+    case raw: Query.Raw =>
+      val prepared = raw.pairs.filter(f)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
+  }
+
+  def filterNot(f: KeyValue => Boolean): Query = this match {
+    case Query.Empty => this
+
+    case parsed: Query.Parsed =>
+      val prepared = parsed.pairs.filterNot(f)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
+
+    case raw: Query.Raw =>
+      val prepared = raw.pairs.filterNot(f)
+      if (prepared.sizeIs == 0) Query.Empty
+      else new Query.Parsed(prepared)
+  }
+
+  def foreach(f: KeyValue => Unit): Unit = this match {
+    case Query.Empty => ()
+    case parsed: Query.Parsed => parsed.pairs.foreach(f)
+    case raw: Query.Raw => raw.pairs.foreach(f)
+  }
+
+  def foldLeft[Z](z: Z)(f: (Z, KeyValue) => Z): Z = this match {
+    case Query.Empty => z
+    case parsed: Query.Parsed => parsed.pairs.foldLeft(z)(f)
+    case raw: Query.Raw => raw.pairs.foldLeft(z)(f)
+  }
 
   def foldRight[Z](z: Eval[Z])(f: (KeyValue, Eval[Z]) => Eval[Z]): Eval[Z] =
-    Foldable[Vector].foldRight(pairs, z)(f)
+    this match {
+      case Query.Empty => z
+      case parsed: Query.Parsed =>
+        Foldable[Vector].foldRight(parsed.pairs, z)(f)
+      case raw: Query.Raw =>
+        Foldable[Vector].foldRight(raw.pairs, z)(f)
+    }
 
-  def +:(elem: KeyValue): Query =
-    new Query(Left(elem +: pairs))
+  def +:(elem: KeyValue): Query = this match {
+    case Query.Empty =>
+      new Query.Parsed(Vector(elem))
+    case parsed: Query.Parsed =>
+      new Query.Parsed(elem +: parsed.pairs)
+    case raw: Query.Raw =>
+      new Query.Parsed(elem +: raw.pairs)
+  }
 
-  def :+(elem: KeyValue): Query =
-    new Query(Left(pairs :+ elem))
+  def :+(elem: KeyValue): Query = this match {
+    case Query.Empty =>
+      new Query.Parsed(Vector(elem))
+    case parsed: Query.Parsed =>
+      new Query.Parsed(parsed.pairs :+ elem)
+    case raw: Query.Raw =>
+      new Query.Parsed(raw.pairs :+ elem)
+  }
 
   def ++(pairs: collection.Iterable[(String, Option[String])]): Query =
-    new Query(Left(this.pairs ++ pairs))
+    this match {
+      case Query.Empty =>
+        new Query.Parsed(pairs.toVector)
+      case parsed: Query.Parsed =>
+        new Query.Parsed(parsed.pairs ++ pairs)
+      case raw: Query.Raw =>
+        new Query.Parsed(raw.pairs ++ pairs)
+    }
 
-  def toVector: Vector[(String, Option[String])] = pairs
+  def toVector: Vector[(String, Option[String])] = this match {
+    case Query.Empty => Vector.empty
+    case parsed: Query.Parsed => parsed.pairs
+    case raw: Query.Raw => raw.pairs
+  }
 
-  def toList: List[(String, Option[String])] = toVector.toList
+  def toList: List[(String, Option[String])] = this match {
+    case Query.Empty => List.empty
+    case parsed: Query.Parsed => parsed.pairs.toList
+    case raw: Query.Raw => raw.pairs.toList
+  }
 
-  /** Render the Query as a `String`.
+  /** Map[String, String] representation of the [[Query]]
     *
-    * Pairs are separated by '&' and keys are separated from values by '='
+    * If multiple values exist for a key, the first is returned. If
+    * none exist, the empty `String` "" is returned.
     */
-  override def render(writer: Writer): writer.type = value.fold(
-    { pairs =>
+  lazy val params: Map[String, String] = this match {
+    case Query.Empty => Map.empty
+    case _: Query.Parsed | _: Query.Raw =>
+      multiParams.map { case (k, v) =>
+        k -> v.headOption.getOrElse("")
+      }
+  }
+
+  /** `Map[String, List[String]]` representation of the [[Query]]
+    *
+    * Params are represented as a `List[String]` and may be empty.
+    */
+  lazy val multiParams: Map[String, List[String]] = this match {
+    case Query.Empty => Map.empty
+    case _: Query.Parsed | _: Query.Raw =>
+      val pairs = toVector
+
+      if (pairs.isEmpty) Map.empty
+      else {
+        val m = mutable.Map.empty[String, ListBuffer[String]]
+        pairs.foreach {
+          case (k, None) => m.getOrElseUpdate(k, new ListBuffer)
+          case (k, Some(v)) => m.getOrElseUpdate(k, new ListBuffer) += v
+        }
+        m.view.mapValues(_.toList).toMap
+      }
+  }
+
+  override protected type Self = Query
+  override protected val query: Query = this
+  override protected def self: Self = this
+  override protected def replaceQuery(query: Query): Self = query
+}
+
+object Query {
+  case object Empty extends Query {
+    def pairs: Vector[KeyValue] = Vector.empty
+
+    override def render(writer: Writer): writer.type =
+      writer
+  }
+
+  final class Raw private[http4s] (value: String) extends Query {
+    private[this] var _pairs: Vector[KeyValue] = _
+
+    def pairs: Vector[KeyValue] = {
+      if (_pairs == null) {
+        _pairs = Query.parse(value)
+      }
+      _pairs
+    }
+
+    override def render(writer: Writer): writer.type =
+      writer.append(value)
+
+    override def equals(that: Any): Boolean =
+      that match {
+        case that: Query => that.toVector == toVector
+        case _ => false
+      }
+
+    override def hashCode: Int = 31 + pairs.##
+  }
+
+  final class Parsed private[http4s] (value: Vector[KeyValue]) extends Query {
+    def pairs: Vector[KeyValue] = value
+
+    /** Render the Query as a `String`.
+      *
+      * Pairs are separated by '&' and keys are separated from values by '='
+      */
+    override def render(writer: Writer): writer.type = {
       var first = true
 
       def encode(s: String) =
@@ -121,7 +295,7 @@ final class Query private (value: Either[Vector[KeyValue], String])
           toSkip = UriCoding.QueryNoEncode,
         )
 
-      pairs.foreach {
+      value.foreach {
         case (n, None) =>
           if (!first) writer.append('&')
           else first = false
@@ -136,79 +310,48 @@ final class Query private (value: Either[Vector[KeyValue], String])
             .append(encode(v))
       }
       writer
-    },
-    raw => writer.append(raw),
-  )
-
-  /** Map[String, String] representation of the [[Query]]
-    *
-    * If multiple values exist for a key, the first is returned. If
-    * none exist, the empty `String` "" is returned.
-    */
-  lazy val params: Map[String, String] =
-    multiParams.map { case (k, v) =>
-      k -> v.headOption.getOrElse("")
     }
 
-  /** `Map[String, List[String]]` representation of the [[Query]]
-    *
-    * Params are represented as a `List[String]` and may be empty.
-    */
-  lazy val multiParams: Map[String, List[String]] =
-    if (toVector.isEmpty) Map.empty
-    else {
-      val m = mutable.Map.empty[String, ListBuffer[String]]
-      toVector.foreach {
-        case (k, None) => m.getOrElseUpdate(k, new ListBuffer)
-        case (k, Some(v)) => m.getOrElseUpdate(k, new ListBuffer) += v
+    override def equals(that: Any): Boolean =
+      that match {
+        case that: Query => that.toVector == toVector
+        case _ => false
       }
-      m.view.mapValues(_.toList).toMap
-    }
 
-  override def equals(that: Any): Boolean =
-    that match {
-      case that: Query => that.toVector == toVector
-      case _ => false
-    }
+    override def hashCode: Int = 31 + value.##
+  }
 
-  override def hashCode: Int = 31 + toVector.##
-
-  // ///////////////////// QueryOps methods and types /////////////////////////
-  override protected type Self = Query
-  override protected val query: Query = this
-  override protected def self: Self = this
-  override protected def replaceQuery(query: Query): Self = query
-  // //////////////////////////////////////////////////////////////////////////
-}
-
-object Query {
   type KeyValue = (String, Option[String])
 
   /** Represents the absence of a query string. */
-  val empty: Query = new Query(Vector.empty)
+  val empty: Query = Query.Empty
 
   /** Represents a query string with no keys or values: `?` */
-  val blank = new Query(Vector("" -> None))
+  val blank = new Query.Parsed(Vector("" -> None))
 
   def apply(xs: (String, Option[String])*): Query =
-    new Query(xs.toVector)
+    if (xs.sizeIs == 0) Query.Empty
+    else new Query.Parsed(xs.toVector)
 
   def fromVector(xs: Vector[(String, Option[String])]): Query =
-    new Query(xs)
+    if (xs.sizeIs == 0) Query.Empty
+    else new Query.Parsed(xs)
 
   def fromPairs(xs: (String, String)*): Query =
-    new Query(
-      xs.toList.foldLeft(Vector.empty[KeyValue]) { case (m, (k, s)) =>
-        m :+ (k -> Some(s))
-      }
-    )
+    if (xs.sizeIs == 0) Query.Empty
+    else
+      new Query.Parsed(
+        xs.toList.foldLeft(Vector.empty[KeyValue]) { case (m, (k, s)) =>
+          m :+ (k -> Some(s))
+        }
+      )
 
   /** Generate a [[Query]] from its `String` representation
     *
     * If parsing fails, the empty [[Query]] is returned
     */
   def unsafeFromString(query: String): Query =
-    if (query.isEmpty) new Query(Vector("" -> None))
+    if (query.isEmpty) new Query.Parsed(Vector("" -> None))
     else
       QueryParser.parseQueryString(query) match {
         case Right(query) => query
@@ -221,7 +364,7 @@ object Query {
 
   /** Build a [[Query]] from the `Map` structure */
   def fromMap(map: collection.Map[String, collection.Seq[String]]): Query =
-    new Query(map.foldLeft(Vector.empty[KeyValue]) {
+    new Query.Parsed(map.foldLeft(Vector.empty[KeyValue]) {
       case (m, (k, Seq())) => m :+ (k -> None)
       case (m, (k, vs)) => vs.toList.foldLeft(m) { case (m, v) => m :+ (k -> Some(v)) }
     })
@@ -244,7 +387,7 @@ object Query {
     import cats.parse.Parser.charIn
     import Rfc3986.pchar
 
-    pchar.orElse(charIn("/?[]")).rep0.string.map(pchars => new Query(Right(pchars)))
+    pchar.orElse(charIn("/?[]")).rep0.string.map(pchars => new Query.Raw(pchars))
   }
 
   implicit val catsInstancesForHttp4sQuery: Hash[Query] with Order[Query] with Show[Query] =