This PR aims to make a model of `Query` more straightforward (and to retire the allocation of `Either`).
```scala
final class Query private (value: Either[Vector[KeyValue], String])
```
versus
```scala
sealed trait Query extends QueryOps with Renderable
object Query {
  case object Empty extends Query 
  final class Raw private[http4s] (value: String) extends Query
  final class Parsed private[http4s] (value: Vector[KeyValue]) extends Query
}
```

It's a proposal, so I'm open to any comments on that.

## Performance
Performance within fluctuation remained the same.

before
```
[info] Benchmark                        (size)   Mode  Cnt       Score       Error   Units
[info] QueryBench.filterParsedQuery          0  thrpt    5  101507.478 ±   419.039  ops/ms
[info] QueryBench.filterParsedQuery         10  thrpt    5   16419.723 ±   145.474  ops/ms
[info] QueryBench.filterParsedQuery        100  thrpt    5    1702.907 ±     9.961  ops/ms
[info] QueryBench.filterParsedQuery       1000  thrpt    5     146.354 ±     0.419  ops/ms
[info] QueryBench.filterRawQuery             0  thrpt    5   92902.523 ±   595.655  ops/ms
[info] QueryBench.filterRawQuery            10  thrpt    5   12797.116 ±  1561.480  ops/ms
[info] QueryBench.filterRawQuery           100  thrpt    5    1376.502 ±    54.955  ops/ms
[info] QueryBench.filterRawQuery          1000  thrpt    5     147.778 ±     2.300  ops/ms
[info] QueryBench.isEmptyParsedQuery         0  thrpt    5  451307.200 ± 26229.110  ops/ms
[info] QueryBench.isEmptyParsedQuery        10  thrpt    5  459002.947 ±  2871.149  ops/ms
[info] QueryBench.isEmptyParsedQuery       100  thrpt    5  463494.037 ±  6096.403  ops/ms
[info] QueryBench.isEmptyParsedQuery      1000  thrpt    5  457077.897 ± 53214.980  ops/ms
[info] QueryBench.isEmptyRawQuery            0  thrpt    5  457737.264 ±  1468.678  ops/ms
[info] QueryBench.isEmptyRawQuery           10  thrpt    5  437616.912 ±  8443.496  ops/ms
[info] QueryBench.isEmptyRawQuery          100  thrpt    5  465647.184 ±  2795.944  ops/ms
[info] QueryBench.isEmptyRawQuery         1000  thrpt    5  460094.743 ± 10366.918  ops/ms
[info] QueryBench.parseQuery                 0  thrpt    5    2715.890 ±    10.020  ops/ms
[info] QueryBench.parseQuery                10  thrpt    5     215.392 ±    28.554  ops/ms
[info] QueryBench.parseQuery               100  thrpt    5      21.097 ±     0.229  ops/ms
[info] QueryBench.parseQuery              1000  thrpt    5       1.907 ±     0.030  ops/ms
[info] QueryBench.prependParsedQuery         0  thrpt    5    1332.830 ±    40.513  ops/ms
[info] QueryBench.prependParsedQuery        10  thrpt    5    1381.962 ±    85.910  ops/ms
[info] QueryBench.prependParsedQuery       100  thrpt    5    1149.776 ±   130.352  ops/ms
[info] QueryBench.prependParsedQuery      1000  thrpt    5    1190.669 ±    86.394  ops/ms
[info] QueryBench.prependRawQuery            0  thrpt    5    1391.598 ±    38.045  ops/ms
[info] QueryBench.prependRawQuery           10  thrpt    5    1381.388 ±    91.171  ops/ms
[info] QueryBench.prependRawQuery          100  thrpt    5    1228.547 ±    45.970  ops/ms
[info] QueryBench.prependRawQuery         1000  thrpt    5    1164.858 ±    90.548  ops/ms
[info] QueryBench.sliceHalfParsedQuery       0  thrpt    5  173791.650 ±  1914.851  ops/ms
[info] QueryBench.sliceHalfParsedQuery      10  thrpt    5   92678.610 ±   666.011  ops/ms
[info] QueryBench.sliceHalfParsedQuery     100  thrpt    5   34289.443 ±  1378.834  ops/ms
[info] QueryBench.sliceHalfParsedQuery    1000  thrpt    5   30999.891 ±    54.828  ops/ms
[info] QueryBench.sliceHalfRawQuery          0  thrpt    5  174246.506 ±  1383.400  ops/ms
[info] QueryBench.sliceHalfRawQuery         10  thrpt    5   94638.441 ±  1246.186  ops/ms
[info] QueryBench.sliceHalfRawQuery        100  thrpt    5   34010.746 ±  2705.160  ops/ms
[info] QueryBench.sliceHalfRawQuery       1000  thrpt    5   29070.040 ±  6898.233  ops/ms
```
after
```
[info] Benchmark                        (size)   Mode  Cnt       Score       Error   Units
[info] QueryBench.filterParsedQuery          0  thrpt    5  125112.052 ±   269.180  ops/ms
[info] QueryBench.filterParsedQuery         10  thrpt    5   18170.039 ±   117.447  ops/ms
[info] QueryBench.filterParsedQuery        100  thrpt    5    1691.472 ±    48.133  ops/ms
[info] QueryBench.filterParsedQuery       1000  thrpt    5     146.326 ±     1.672  ops/ms
[info] QueryBench.filterRawQuery             0  thrpt    5  119818.520 ±  3184.840  ops/ms
[info] QueryBench.filterRawQuery            10  thrpt    5   13899.793 ±    63.296  ops/ms
[info] QueryBench.filterRawQuery           100  thrpt    5    1545.466 ±     1.546  ops/ms
[info] QueryBench.filterRawQuery          1000  thrpt    5     151.913 ±     2.996  ops/ms
[info] QueryBench.isEmptyParsedQuery         0  thrpt    5  438149.754 ± 22818.420  ops/ms
[info] QueryBench.isEmptyParsedQuery        10  thrpt    5  447012.329 ±  4631.002  ops/ms
[info] QueryBench.isEmptyParsedQuery       100  thrpt    5  462690.306 ±  4388.934  ops/ms
[info] QueryBench.isEmptyParsedQuery      1000  thrpt    5  462609.103 ±  5139.394  ops/ms
[info] QueryBench.isEmptyRawQuery            0  thrpt    5  447738.371 ±   594.162  ops/ms
[info] QueryBench.isEmptyRawQuery           10  thrpt    5  434970.727 ± 12788.456  ops/ms
[info] QueryBench.isEmptyRawQuery          100  thrpt    5  451769.683 ± 19311.524  ops/ms
[info] QueryBench.isEmptyRawQuery         1000  thrpt    5  464884.523 ±  1643.294  ops/ms
[info] QueryBench.parseQuery                 0  thrpt    5    2718.778 ±    32.716  ops/ms
[info] QueryBench.parseQuery                10  thrpt    5     216.652 ±     0.846  ops/ms
[info] QueryBench.parseQuery               100  thrpt    5      20.982 ±     0.113  ops/ms
[info] QueryBench.parseQuery              1000  thrpt    5       1.914 ±     0.016  ops/ms
[info] QueryBench.prependParsedQuery         0  thrpt    5    1300.867 ±    34.019  ops/ms
[info] QueryBench.prependParsedQuery        10  thrpt    5    1522.673 ±    21.713  ops/ms
[info] QueryBench.prependParsedQuery       100  thrpt    5    1176.870 ±    46.216  ops/ms
[info] QueryBench.prependParsedQuery      1000  thrpt    5    1202.466 ±    78.854  ops/ms
[info] QueryBench.prependRawQuery            0  thrpt    5    1424.427 ±   266.030  ops/ms
[info] QueryBench.prependRawQuery           10  thrpt    5    1490.961 ±    82.051  ops/ms
[info] QueryBench.prependRawQuery          100  thrpt    5    1276.975 ±   286.531  ops/ms
[info] QueryBench.prependRawQuery         1000  thrpt    5    1258.148 ±   230.296  ops/ms
[info] QueryBench.sliceHalfParsedQuery       0  thrpt    5  168781.553 ±   825.006  ops/ms
[info] QueryBench.sliceHalfParsedQuery      10  thrpt    5   92949.332 ±  1552.694  ops/ms
[info] QueryBench.sliceHalfParsedQuery     100  thrpt    5   39651.132 ±   200.127  ops/ms
[info] QueryBench.sliceHalfParsedQuery    1000  thrpt    5   31954.834 ±    94.616  ops/ms
[info] QueryBench.sliceHalfRawQuery          0  thrpt    5  170116.904 ±   246.680  ops/ms
[info] QueryBench.sliceHalfRawQuery         10  thrpt    5   93275.785 ±   318.445  ops/ms
[info] QueryBench.sliceHalfRawQuery        100  thrpt    5   39352.343 ±    55.206  ops/ms
[info] QueryBench.sliceHalfRawQuery       1000  thrpt    5   31805.869 ±   139.338  ops/ms
```