@danicheg 2022-12-04
Thanks for the review, folks. But let's wait for a while for comments from other maintainers since this heavily changes the model of `Query`. Yet another week should be enough, I suppose.